+++ front/src/components/BaseRadio.vue
@@ -17,7 +17,7 @@
   </div>
 </template>
 <script>
-import { randomString } from "./stringUtils";
+import { randomString } from "../utils/stringUtils";
 export default {
   name: "base-radio",
   props: {
