+++ front/src/utils/stringUtils.js
