+++ front/src/views/SeriesProfile.vue
@@ -91,7 +91,7 @@
   </div>
 </template>
 <script>
-import { isUser, isAdmin } from "../components/authUtils";
+import { isUser, isAdmin } from "../utils/authUtils";
 
 export default {
   data() {
