+++ front/src/views/StudioEdit.vue
@@ -57,7 +57,7 @@
   </div>
 </template>
 <script>
-import { isAdmin } from "../components/authUtils";
+import { isAdmin } from "../utils/authUtils";
 export default {
   name: "user-profile",
   data() {
