+++ front/src/views/StudioProfile.vue
@@ -92,7 +92,7 @@
   </div>
 </template>
 <script>
-import { isUser, isAdmin } from "../components/authUtils";
+import { isUser, isAdmin } from "../utils/authUtils";
 
 export default {
   name: "studio-profile",
